Add new class and refactor classes
There's one more new class called RandomStringGenerator. There's also several classes with the package correctly put in the class
