+++ src/es/deloitte/git/v3/WeeklyAlgorithm.java
@@ -1,4 +1,4 @@
-package es.deloitte.git;
+package es.deloitte.git.v3;
 
 import java.util.Arrays;
 import java.util.Comparator;
