+++ src/es/deloitte/git/v4/WeeklyAlgorithm.java
@@ -1,4 +1,4 @@
-package es.deloitte.git.v2;
+package es.deloitte.git.v4;
 
 import java.util.Arrays;
 import java.util.Comparator;
