+++ src/es/deloitte/git/v5/WeeklyAlgorithm.java
@@ -1,4 +1,4 @@
-package es.deloitte.git.v2;
+package es.deloitte.git.v5;
 
 import java.util.Arrays;
 import java.util.Comparator;
